IntakeAutoGrabBalls Added
This command takes in a specified time (duration of the movement of the drive), and will run the shooter two seconds before that period ends. This way our shooter motors will be revved and warm when we are ready to shoot. Use PathPlanner to find time taken to move each portion